requirements.txt
torch
torchvision
openai==0.28.0
pillow
matplotlib
